+++ pic2vec/enums.py
@@ -1,0 +1,9 @@
+"""
+This file contains a list of enums that are used across the entire pic2vec
+package.
+"""
+# List of models supported in pic2vec
+MODELS = ['squeezenet', 'vgg16', 'vgg19', 'resnet50', 'inceptionv3', 'xception']
+
+# Tolerance for prediction error
+ATOL = 0.00001